moved all little filter logic into map type plugins and specialised the marker-config functions within each plugin
