+++ packages/libs/eda/src/lib/map/analysis/mapTypes/plugins/DonutMarkerMapType.tsx
@@ -39,10 +39,10 @@ import {
   useDistributionOverlayConfig,
   isNoDataError,
   noDataErrorMessage,
-  visibleOptionFilterTypes,
-  markerDataFilterTypes,
-  floaterFilterTypes,
-  TimeSliderComponent,
+  visibleOptionFilterFuncs,
+  markerDataFilterFuncs,
+  floaterFilterFuncs,
+  pieOrBarMarkerConfigLittleFilter,
 } from '../shared';
 import {
   MapTypeConfigPanelProps,
@@ -62,6 +62,7 @@ import Spinner from '@veupathdb/components/lib/components/Spinner';
 import { MapFloatingErrorDiv } from '../../MapFloatingErrorDiv';
 import { MapTypeHeaderCounts } from '../MapTypeHeaderCounts';
 import { useLittleFilters } from '../../littleFilters';
+import TimeSliderQuickFilter from '../../TimeSliderQuickFilter';
 
 const displayName = 'Donuts';
 
@@ -103,12 +104,14 @@ function ConfigPanelComponent(props: MapTypeConfigPanelProps) {
     );
   }
 
-  const { filters: filtersForVisibleOption } = useLittleFilters({
-    filters,
-    appState,
-    geoConfigs,
-    filterTypes: visibleOptionFilterTypes,
-  });
+  const { filters: filtersForVisibleOption } = useLittleFilters(
+    {
+      filters,
+      appState,
+      geoConfigs,
+    },
+    visibleOptionFilterFuncs
+  );
 
   const allFilteredCategoricalValues = useCategoricalValues({
     overlayEntity,
@@ -291,12 +294,14 @@ function MapLayerComponent(props: MapTypeMapLayerProps) {
   const { selectedVariable, binningMethod, selectedValues } =
     props.configuration as PieMarkerConfiguration;
 
-  const { filters: filtersForMarkerData } = useLittleFilters({
-    filters,
-    appState,
-    geoConfigs,
-    filterTypes: markerDataFilterTypes,
-  });
+  const { filters: filtersForMarkerData } = useLittleFilters(
+    {
+      filters,
+      appState,
+      geoConfigs,
+    },
+    markerDataFilterFuncs
+  );
 
   const markerDataResponse = useMarkerData({
     studyId,
@@ -369,12 +374,14 @@ function MapOverlayComponent(props: MapTypeMapLayerProps) {
     [props.configuration, updateConfiguration]
   );
 
-  const { filters: filtersForFloaters } = useLittleFilters({
-    filters,
-    appState,
-    geoConfigs,
-    filterTypes: floaterFilterTypes,
-  });
+  const { filters: filtersForFloaters } = useLittleFilters(
+    {
+      filters,
+      appState,
+      geoConfigs,
+    },
+    floaterFilterFuncs
+  );
 
   const data = useMarkerData({
     studyId,
@@ -446,12 +453,14 @@ function MapTypeHeaderDetails(props: MapTypeMapLayerProps) {
   const { selectedVariable, binningMethod, selectedValues } =
     props.configuration as PieMarkerConfiguration;
 
-  const { filters: filtersForMarkerData } = useLittleFilters({
-    filters,
-    appState,
-    geoConfigs,
-    filterTypes: markerDataFilterTypes,
-  });
+  const { filters: filtersForMarkerData } = useLittleFilters(
+    {
+      filters,
+      appState,
+      geoConfigs,
+    },
+    markerDataFilterFuncs
+  );
 
   const markerDataResponse = useMarkerData({
     studyId,
@@ -478,6 +487,52 @@ function MapTypeHeaderDetails(props: MapTypeMapLayerProps) {
     />
   ) : null;
 }
+
+const timeSliderFilterFuncs = [pieOrBarMarkerConfigLittleFilter];
+
+export function TimeSliderComponent(props: MapTypeMapLayerProps) {
+  const {
+    studyId,
+    studyEntities,
+    filters,
+    appState,
+    appState: { timeSliderConfig },
+    analysisState,
+    geoConfigs,
+    setTimeSliderConfig,
+    siteInformationProps,
+  } = props;
+
+  const toggleStarredVariable = useToggleStarredVariable(analysisState);
+  const findEntityAndVariable = useFindEntityAndVariable(filters);
+
+  const { filters: filtersForTimeSlider } = useLittleFilters(
+    {
+      filters,
+      appState,
+      geoConfigs,
+      findEntityAndVariable,
+    },
+    timeSliderFilterFuncs
+  );
+
+  return timeSliderConfig && setTimeSliderConfig && siteInformationProps ? (
+    <TimeSliderQuickFilter
+      studyId={studyId}
+      entities={studyEntities}
+      filters={filtersForTimeSlider}
+      starredVariables={
+        analysisState.analysis?.descriptor.starredVariables ?? []
+      }
+      toggleStarredVariable={toggleStarredVariable}
+      config={timeSliderConfig}
+      updateConfig={setTimeSliderConfig}
+      siteInformation={siteInformationProps}
+    />
+  ) : null;
+}
+
+////// functions and hooks ///////
 
 function useMarkerData(props: DistributionMarkerDataProps) {
   const {
